bitcoin_project/bitcoin_prediction/requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
Flask==2.1.2
fonttools==4.33.3
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
plotly==5.8.0
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
tenacity==8.0.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
